[shallowequal] Update types to v1.1 (#31765)
* [shallowequal] Move v0 type to sub-dir

* [shallowequal] Update types to v1.1
